[NCL-5701] - Port tests to RestResponseFormattingTest
