+++ rest-new/src/main/java/org/jboss/pnc/rest/JaxRsActivatorNew.java
@@ -50,6 +50,7 @@ import org.jboss.pnc.rest.endpoints.internal.BpmEndpointImpl;
 import org.jboss.pnc.rest.endpoints.internal.BuildMaintenanceEndpointImpl;
 import org.jboss.pnc.rest.endpoints.internal.BuildTaskEndpointImpl;
 import org.jboss.pnc.rest.endpoints.internal.CacheEndpointImpl;
+import org.jboss.pnc.rest.endpoints.internal.DebugEndpointImpl;
 import org.jboss.pnc.rest.endpoints.internal.GenericSettingEndpointImpl;
 import org.jboss.pnc.rest.endpoints.internal.HealthCheckEndpointImpl;
 import org.jboss.pnc.rest.jackson.JacksonProvider;
@@ -211,6 +212,7 @@ public class JaxRsActivatorNew extends Application {
         resources.add(HealthCheckEndpointImpl.class);
         resources.add(GenericSettingEndpointImpl.class);
         resources.add(CacheEndpointImpl.class);
+        resources.add(DebugEndpointImpl.class);
     }
 
     private void addExceptionMappers(Set<Class<?>> resources) {
